moved CreateTable component to layout folder
