+++ front-end/src/layout/CreateTable.js
@@ -1,5 +1,5 @@
 import React from "react";
-import TableForm from "./TableForm";
+import TableForm from "../table/TableForm";
 
 export default function CreateTable() {
     return (
